Update discovery service data model

+++ src/main/java/eu/opertusmundi/common/model/discovery/client/ClientJoinableTableMatchDto.java
@@ -50,7 +50,7 @@ public class ClientJoinableTableMatchDto {
     @Setter
     public static class ClientJoinableTableMatchRelatedDto {
 
-        private Integer coma;
+        private Double coma;
 
         public static ClientJoinableTableMatchRelatedDto from(ServerJoinableTableMatchRelatedDto s) {
             final var c = new ClientJoinableTableMatchRelatedDto();
